feat: replicates openMSX behavior where LDIR waits for BC to become 0
The Z80 executes the LDIR instruction as: execute LDI, then if BC is not 
equal to zero repeat the same instruction. With 'repeat' as in "do not 
increment PC". This way the Z80 will re-fetch the same LDIR instruction 
from memory and re-execute it. One important aspect of this 
implementation is that it allows to serve interrupt request during 
execution of a (long) LDIR instruction.

OpenMSX emulates the LDIR instruction in this way. That means that 
"debug step" on an LDIR instruction will only advance past this 
instruction if BC has become zero.

https://www.msx.org/forum/msx-talk/openmsx/ldir-takes-two-step-over-commands-to-go-to-next-instruction